+++ eclipse-plugin-studio/src/com/twinsoft/convertigo/eclipse/views/projectexplorer/model/DatabaseObjectTreeObject.java
@@ -156,6 +156,7 @@ public class DatabaseObjectTreeObject extends TreeParent implements TreeObjectLi
 		super(viewer, object);
 		isInherited = inherited;
 		hasBeenModified((object.bNew) || (object.hasChanged && !object.bNew));
+		introspect();
 	}
 
 	@Override
@@ -241,13 +242,29 @@ public class DatabaseObjectTreeObject extends TreeParent implements TreeObjectLi
 
 	protected void reloadDescriptors() {
 		propertyDescriptors = null;
-		getDescriptors();
 	}
 
 	protected List<PropertyDescriptor> getDynamicPropertyDescriptors() {
 		return new ArrayList<PropertyDescriptor>();
 	}
 
+	private boolean introspect() {
+		if (databaseObjectBeanDescriptor != null && databaseObjectPropertyDescriptors != null) {
+			return true;
+		}
+		DatabaseObject databaseObject = getObject();
+		try {
+			BeanInfo bi = databaseObjectBeanInfo = CachedIntrospector.getBeanInfo(databaseObject.getClass());
+			databaseObjectBeanDescriptor = bi.getBeanDescriptor();
+			databaseObjectPropertyDescriptors = bi.getPropertyDescriptors();
+			return true;
+		}
+		catch (Exception e) {
+			String message = "Error while introspecting object " + databaseObject.getName() + " (" + databaseObject.getQName() + ")";
+			ConvertigoPlugin.logException(e, message);
+			return false;
+		}
+	}
 
 	protected synchronized void getDescriptors() {
 		if (propertyDescriptors != null && databaseObjectBeanDescriptor != null &&
@@ -262,21 +279,13 @@ public class DatabaseObjectTreeObject extends TreeParent implements TreeObjectLi
 		if ((!(databaseObject instanceof Project)) && (databaseObject.getParent() == null))
 			return; // No needs for removed object
 
-		int len;
-
-		java.beans.PropertyDescriptor databaseObjectPropertyDescriptor;
-
-		try {
-			BeanInfo bi = databaseObjectBeanInfo = CachedIntrospector.getBeanInfo(databaseObject.getClass());
-			databaseObjectBeanDescriptor = bi.getBeanDescriptor();
-			databaseObjectPropertyDescriptors = bi.getPropertyDescriptors();
-			len = databaseObjectPropertyDescriptors.length;
-		}
-		catch (Exception e) {
-			String message = "Error while introspecting object " + databaseObject.getName() + " (" + databaseObject.getQName() + ")";
-			ConvertigoPlugin.logException(e, message);
+		if (!introspect()) {
 			return;
 		}
+
+		int len = databaseObjectPropertyDescriptors.length;
+
+		java.beans.PropertyDescriptor databaseObjectPropertyDescriptor;
 
 		List<PropertyDescriptor> vPropertyDescriptors = new ArrayList<PropertyDescriptor>(32);
 
@@ -675,6 +684,7 @@ public class DatabaseObjectTreeObject extends TreeParent implements TreeObjectLi
 
 	private PropertyDescriptor findPropertyDescriptor(Object id) {
 		PropertyDescriptor propertyDescriptor;
+		getDescriptors();
 		if (propertyDescriptors != null) {
 			for (int i=0;i<propertyDescriptors.length;i++) {
 				propertyDescriptor = (PropertyDescriptor)propertyDescriptors[i];
@@ -1266,9 +1276,6 @@ public class DatabaseObjectTreeObject extends TreeParent implements TreeObjectLi
 	public void treeObjectAdded(TreeObjectEvent treeObjectEvent) {
 		checkDone(treeObjectEvent);
 		checkReset(treeObjectEvent);
-		DatabaseObjectTreeObject treeObject = (DatabaseObjectTreeObject)treeObjectEvent.getSource();
-		if (!(treeObject.equals(this)))
-			getDescriptors();// refresh editors (e.g labels in combobox)
 	}
 
 	public void treeObjectRemoved(TreeObjectEvent treeObjectEvent) {
@@ -1302,9 +1309,6 @@ public class DatabaseObjectTreeObject extends TreeParent implements TreeObjectLi
 				}
 			}
 		}
-
-		if (!(treeObject.equals(this)) && (!getParents().contains(treeObject)))
-			getDescriptors();// refresh editors (e.g labels in combobox)
 	}
 
 	public void treeObjectPropertyChanged(TreeObjectEvent treeObjectEvent) {
@@ -1351,8 +1355,6 @@ public class DatabaseObjectTreeObject extends TreeParent implements TreeObjectLi
 				}
 			}
 		}
-
-		getDescriptors();// refresh editors (e.g labels in combobox)
 	}
 
 	public boolean hasAncestorDisabled(){